feat: applicant & collection validation
